webapp/requirements.txt
requests~=2.31.0
mysql-connector-python~=9.0.0
flask~=2.2.2